11/2/2023 LSTM
Here lies the LSTM python script that I used to do 30 runs of LSTM using by_day missing person data. I include the txt files of all results. As well, I include the R script (rmd) I used to make graphs for it.

There's also 2 pdfs. One is all data from these results so you can compare them with your eyes. Including error tables and graphs. The second one is a pdf I wrote a while ago to analyze previous models used, but I never uploaded it. So here it is.

There is also the html attached that is generated by the rmd or R markdown file that was used to make all the graphs.